Adding a StandardsRegExt record.
This record should be sent to the RofR once this is merged.

I've updated the DOCDATE, which is reflected in the StandardsRegExt record,
so it'd be great if you could keep it (or fix it in both locations).

This also updates ivoatex to the current git status (in the future,
run make updated now and then).

Finally, I've taken out the role diagram from the vectorfigures, as
its PNG isn't necessary any more (the HTML uses the SVG version).